# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PubWIFIprot API",
        "description": "API documentation for the PubWIFIprot backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Development Server"
        },
        {
            "url": "https://api.pubwifiprot.com",
            "description": "Production Server"
        }
    ],
    "paths": {
...
        "/api/sessions/start": {
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Start a new session",
                "operationId": "eaca49322cdf04b3f77755b3f7de3586",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "device_id",
                                    "network_ssid",
                                    "network_bssid"
                                ],
                                "properties": {
                                    "device_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "network_ssid": {
                                        "type": "string",
                                        "example": "CoffeeShop_WiFi"
                                    },
                                    "network_bssid": {
                                        "type": "string",
                                        "example": "00:11:22:33:44:55"
                                    },
                                    "location": {
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": 40.7128
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -74.006
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session started successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Session started successfully"
                                        },
                                        "session": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sessions/end": {
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "End an active session",
                "operationId": "35e59eae6c26c1e101da01d9a6fc209e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "summary": {
                                        "properties": {
                                            "total_dns_requests": {
                                                "type": "integer",
                                                "example": 150
                                            },
                                            "blocked_domains": {
                                                "type": "integer",
                                                "example": 3
                                            },
                                            "network_changes": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session ended successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Session ended successfully"
                                        },
                                        "session": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sessions/update": {
            "post": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Update an active session with new activity",
                "operationId": "00fbde8e30c71dce8e3ca864e087648f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "session_id",
                                    "activity"
                                ],
                                "properties": {
                                    "session_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "activity": {
                                        "properties": {
                                            "type": {
                                                "type": "string",
                                                "enum": [
                                                    "dns_request",
                                                    "network_change",
                                                    "blocked_domain"
                                                ]
                                            },
                                            "details": {
                                                "type": "object"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Session activity updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Session activity updated successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/sessions/history": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "summary": "Get session history for the authenticated user",
                "operationId": "737608475502b91715d0421835c98a28",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of user sessions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "device_id": {
                                                        "type": "integer"
                                                    },
                                                    "network_ssid": {
                                                        "type": "string"
                                                    },
                                                    "network_bssid": {
                                                        "type": "string"
                                                    },
                                                    "location": {
                                                        "type": "object"
                                                    },
                                                    "started_at": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "ended_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "nullable": true
                                                    },
                                                    "summary": {
                                                        "type": "object",
                                                        "nullable": true
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your JWT token in the format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}